*(ticdc): move memory sorter
